get rid of utils and move countries to it's own package
