+++ countries/alpha-2-to-alpha-3-lookup.go
@@ -1,4 +1,4 @@
-package country_codes
+package countries
 
 var Alpha2ToAlpha3 = map[string]string{
 	"BD": "BGD",
