+++ countries/iso-3166-1-alpha-3-codes.go
@@ -1,4 +1,4 @@
-package country_codes
+package countries
 
 type Alpha3CountryCodeDetails struct {
 	CountryCode string
