+++ routes/admin_jumio.go
@@ -5,7 +5,7 @@ import (
 	"encoding/gob"
 	"encoding/json"
 	"fmt"
-	"github.com/deso-protocol/backend/utils/country_codes"
+	"github.com/deso-protocol/backend/countries"
 	"github.com/deso-protocol/core/lib"
 	"github.com/golang/glog"
 	"io"
@@ -234,7 +234,7 @@ func (fes *APIServer) AdminJumioCallback(ww http.ResponseWriter, req *http.Reque
 		return
 	}
 
-	if _, exists := country_codes.Alpha3CountryCodes[requestData.CountryAlpha3]; requestData.CountryAlpha3 != "" && !exists {
+	if _, exists := countries.Alpha3CountryCodes[requestData.CountryAlpha3]; requestData.CountryAlpha3 != "" && !exists {
 		_AddBadRequestError(ww, fmt.Sprintf("AdminJumioCallback: Invalid country alpha-3: %s", requestData.CountryAlpha3))
 		return
 	}
@@ -265,10 +265,10 @@ func (fes *APIServer) AdminUpdateJumioCountrySignUpBonus(ww http.ResponseWriter,
 			"Problem parsing request body: %v", err))
 		return
 	}
-	var countryCodeDetails country_codes.Alpha3CountryCodeDetails
+	var countryCodeDetails countries.Alpha3CountryCodeDetails
 	var exists bool
 	// Validate the country code
-	if countryCodeDetails, exists = country_codes.Alpha3CountryCodes[requestData.CountryCode]; !exists {
+	if countryCodeDetails, exists = countries.Alpha3CountryCodes[requestData.CountryCode]; !exists {
 		_AddBadRequestError(ww, fmt.Sprintf("AdminUpdateJumioCountrySignUpBonus: "+
 			"invalid country code: %v", requestData.CountryCode))
 		return
@@ -312,13 +312,13 @@ func (fes *APIServer) AdminGetAllCountryLevelSignUpBonuses(ww http.ResponseWrite
 
 type CountrySignUpBonusResponse struct {
 	CountryLevelSignUpBonus CountryLevelSignUpBonus
-	CountryCodeDetails      country_codes.Alpha3CountryCodeDetails
+	CountryCodeDetails      countries.Alpha3CountryCodeDetails
 }
 
 // SetAllCountrySignUpBonusMetadata goes through all countries in map of CountryCodes in utils and sets the sign-up
 // bonus config in cache.
 func (fes *APIServer) SetAllCountrySignUpBonusMetadata() {
-	for countryCode, countryDetails := range country_codes.Alpha3CountryCodes {
+	for countryCode, countryDetails := range countries.Alpha3CountryCodes {
 		signUpBonus, err := fes.GetJumioCountrySignUpBonus(countryCode)
 		if err != nil {
 			glog.Errorf("SetAllCountrySignUpBonusMetadata: %v", err)
@@ -330,7 +330,7 @@ func (fes *APIServer) SetAllCountrySignUpBonusMetadata() {
 }
 
 // SetSingleCountrySignUpBonus sets the sign up bonus configuration for a given country in the cached map.
-func (fes *APIServer) SetSingleCountrySignUpBonus(countryDetails country_codes.Alpha3CountryCodeDetails,
+func (fes *APIServer) SetSingleCountrySignUpBonus(countryDetails countries.Alpha3CountryCodeDetails,
 	signUpBonus CountryLevelSignUpBonus) {
 	fes.AllCountryLevelSignUpBonuses[countryDetails.Name] = CountrySignUpBonusResponse{
 		CountryLevelSignUpBonus: signUpBonus,
@@ -341,7 +341,7 @@ func (fes *APIServer) SetSingleCountrySignUpBonus(countryDetails country_codes.A
 // GetSingleCountrySignUpBonus returns the current value of the sign-up bonus configuration stored in the cached map.
 func (fes *APIServer) GetSingleCountrySignUpBonus(countryCode string) CountryLevelSignUpBonus {
 	// Convert country code to uppercase just in case.
-	countryCodeDetails := country_codes.Alpha3CountryCodes[strings.ToUpper(countryCode)]
+	countryCodeDetails := countries.Alpha3CountryCodes[strings.ToUpper(countryCode)]
 	// If we can't find the signup bonus from the map, return the default. Else, return the sign up bonus we found in
 	// the map.
 	if countrySignUpBonusResponse, exists := fes.AllCountryLevelSignUpBonuses[countryCodeDetails.Name]; !exists {
